feat: set notebook package as optional
It allows to avoid an error if `npx now` is run on a fresh copy of the
repository. With that commit, when now.sh runs `npm install`, the
aaa_notebook dependency is ignored. When it then runs `npm build`, the
notebook is downloaded and installed, and the data is downloaded.