fix: ignore yarn cache when updating notebook package
yarn doesn't have any mechanism yet to avoid using the cache when
installing a package (see https://github.com/yarnpkg/yarn/issues/2165).
But for local packages, we want to force install whan we download a new
version.
The trick here is to used a different filename for the tgz file... A
hack, not a solution.